Initial Commit of UFC font for F/A-18C & AV-8B
These are two fonts for the up front controller  on the F/A-18C and AV-8B.   The first one is for the single digit 7 segment displays and the wider font is for the 16 segment displays.  The fonts were build at https://glyphter.com/ and for the wider font, struggled to find appropriate slots for many of the symbols.